+++ src/global.d.ts
@@ -1,0 +1,16 @@
+import type {Linter} from 'eslint';
+import type {PublicApi} from 'stylelint';
+
+interface luaparse {
+	defaultOptions: {luaVersion: string};
+	parse(s: string): void;
+	SyntaxError: new () => {message: string, index: number};
+}
+
+declare global {
+	const eslint: {
+		Linter: new () => Linter;
+	};
+	const stylelint: PublicApi;
+	const luaparse: luaparse;
+}